Issue #202: Reduced the file length of InputSurvivingMutations
# Conflicts:
#	config/checkstyle-resources-suppressions.xml
