+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/survivingmutations/DiffSurvivingMutations.diff
@@ -62,7 +62,7 @@
      }
  
                                 
-@@ -90,55 +85,49 @@
+@@ -90,10 +85,9 @@
      int m5(int x, int y) {
          return switch (x) { // (suppressed violation)
              case 1: // (suppressed violation)
@@ -75,59 +75,3 @@
                  };
              default: yield 0;
          };
-     }
-     
-     void m6(int x) {
--          // violation below, 'Switch can be replaced with enhanced switch'
-           switch (x) {
--            case 1: break;
--            default: break;
-+            case 1 -> {}
-+            default -> {}
-           }
-     }
- 
-     void m7(int x) {
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--case 1:
-+case 1 -> {
- System.out.println("0");
- System.out.println("0b");
--break;
--  default:
-+}
-+  default -> {
-   System.out.println("2");
-   System.out.println("2b");
--  break;
-+  }
-         }
-     }
- 
-     void m8(int x, int y) {
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) { case 1: switch (y) { case 1: System.out.println("1"); case 2: break; default: break; } break;
--            default: break;
-+        switch (x) { case 1 -> { switch (y) { case 1 -> System.out.println("1"); case 2 -> {} default -> {} }}
-+            default -> {}
-         }
-     }
- 
-     int m9(int x, int y) {
--        // violation below, 'Switch can be replaced with enhanced switch'
--        return switch (x) { case 1 -> switch (y) { case 1: System.out.println("1"); yield 1; case 2: yield 2; default: yield 0; };
-+        return switch (x) { case 1 -> switch (y) { case 1 -> { System.out.println("1"); yield 1;} case 2 -> 2; default -> 0; };
-             default -> 0;
-         };
-     }
- 
-     void m10(int x) {
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) { case 1:      System.out.println("1"); System.out.println("1b"); break;
--            default: break;
-+        switch (x) { case 1 -> {      System.out.println("1"); System.out.println("1b");}
-+            default -> {}
-         }
-     }
-
