+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/survivingmutations3/DiffSurvivingMutations3.diff
@@ -1,0 +1,63 @@
+--- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/survivingmutations3/InputSurvivingMutations3.java
++++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/survivingmutations3/OutputSurvivingMutations3.java
+@@ -13,48 +13,43 @@
+ 
+ package org.checkstyle.autofix.recipe.useenhancedswitch.survivingmutations3;
+ 
+-public class InputSurvivingMutations3 {
++public class OutputSurvivingMutations3 {
+     
+     void m6(int x) {
+-          // violation below, 'Switch can be replaced with enhanced switch'
+           switch (x) {
+-            case 1: break;
+-            default: break;
++            case 1 -> {}
++            default -> {}
+           }
+     }
+ 
+     void m7(int x) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+         switch (x) {
+-case 1:
++case 1 -> {
+ System.out.println("0");
+ System.out.println("0b");
+-break;
+-  default:
++}
++  default -> {
+   System.out.println("2");
+   System.out.println("2b");
+-  break;
++  }
+         }
+     }
+ 
+     void m8(int x, int y) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) { case 1: switch (y) { case 1: System.out.println("1"); case 2: break; default: break; } break;
+-            default: break;
++        switch (x) { case 1 -> { switch (y) { case 1 -> System.out.println("1"); case 2 -> {} default -> {} }}
++            default -> {}
+         }
+     }
+ 
+     int m9(int x, int y) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        return switch (x) { case 1 -> switch (y) { case 1: System.out.println("1"); yield 1; case 2: yield 2; default: yield 0; };
++        return switch (x) { case 1 -> switch (y) { case 1 -> { System.out.println("1"); yield 1;} case 2 -> 2; default -> 0; };
+             default -> 0;
+         };
+     }
+ 
+     void m10(int x) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) { case 1:      System.out.println("1"); System.out.println("1b"); break;
+-            default: break;
++        switch (x) { case 1 -> {      System.out.println("1"); System.out.println("1b");}
++            default -> {}
+         }
+     }
+
